// frontend/src/app/model/Client.ts
export class Client {
	client_id: Number;
	clien_numero: Number;
	clien_typeclient: String;
	clien_numerosolvabilite: Number;
	clien_nature: String;
	clien_typesociete: Number;
	clien_coderegroupgestion: Number;
	clien_titre: String;
	clien_nom: String;
	clien_prenom: String;
	clien_denomination: String;
	clien_sigle: String;
	clien_adressenumero: String;
	clien_adresserue: String;
	clien_adresseville: String;
	clien_datenaissance: Date;
	// clien_datesouscription1: Date;
	clien_categsocioprof: Number;
	clien_telephone1: String;
	clien_telephone2: String;
	clien_portable: String;
	// clien_fax: String;
	clien_email: String;
	clien_website: String;
	clien_ninea: String;
	clien_registrecommerce: String;
	clien_codeorigine: String;
	clien_sexe: String;
	clien_formejuridique: Number;
	clien_effectif: Number;
	clien_chiffreaffaireannuel: Number;
	clien_chiffreaffaireprime: Number;
	clien_chargesinistre: Number;
	clien_contactprinci: String;
	clien_utilisateur: Number;
	clien_datemodification: Date;
	clien_anciennumero: Number;
	clien_CIN: String;
	clien_capital_social: Number;
	clien_date_relation: Date;
	clien_secteur_activite: Number;
	clien_status: String;
	clien_princdirigeant: String;
	clien_modegouvernance: String;
	clien_principalactionnaire: String;
	clien_facebook: String;
	clien_linkdin: String;
	clien_activitecommercante: String;
	clien_passeport: String ;
	clien_commentaire: Text ;
	clien_numeroprospect: Number;
}